backend/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dateutil==2.8.2
Werkzeug==2.3.7
email-validator==2.1.0
python-multipart==0.0.6
python-slugify==8.0.1
Pillow>=9.0.0
gunicorn==21.2.0
flask-cors==4.0.0
